requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.7
Flask==2.3.2
Flask-Cors==4.0.0
flask-restx==1.1.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pytz==2023.3
referencing==0.30.2
rpds-py==0.9.2
Werkzeug==2.3.7
zipp==3.16.2
pymysql
python-dotenv
spotipy